+++ src/plone/jsonapi/core/browser/decorators.py
@@ -106,22 +106,6 @@ def returns_json(func):
     return decorator
 
 
-def supports_jsonp(func):
-    """ suports jsonp output
-    """
-
-    def decorator(*args, **kwargs):
-        instance = args[0]
-        request = getattr(instance, "request", None)
-
-        c = request.form.get("c", None)
-        if c is not None:
-            return "%s(%s);" % (str(c), func(*args, **kwargs))
-        return func(*args, **kwargs)
-
-    return decorator
-
-
 def returns_binary_stream(func):
     """ returns a binary file stream
     """
@@ -131,7 +115,8 @@ def returns_binary_stream(func):
         request = getattr(instance, "request", None)
         request.response.setHeader("Content-Type", "application/zip")
         zip_out = func(*args, **kwargs)
-        request.response.setHeader("Content-Length", str(os.path.getsize(zip_out)))
+        request.response.setHeader(
+            "Content-Length", str(os.path.getsize(zip_out)))
         return filestream_iterator(zip_out)
 
     return decorator